Warnings removed
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/core/trunk@497 77d55b08-817d-0410-979b-1fec1beb8c2e

+++ cluster/CommChannelServer.cpp
@@ -146,9 +146,8 @@ void cluster::CommChannelServer::SingleSendMessage(const vislib::net::AbstractSi
 /**
  * Answers the number of clients currently connected.
  */
-unsigned int cluster::CommChannelServer::ClientsCount()
-{
-	return this->clients.Count();
+unsigned int cluster::CommChannelServer::ClientsCount() {
+    return static_cast<unsigned int>(this->clients.Count());
 }
 
 /*
